requirements.txt
langchain
langgraph
langchain-core
langchain_community
langchain_groq
langchain_openai
faiss-cpu
streamlit
tavily-python
python-dotenv
llama-cpp-python
uvicorn[standard]==0.29.0
fastapi==0.115.1
loguru==0.7.2
pytest
websockets
requests
spacy
redis
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl